+++ src/upnp/SoapActionsWANEthernetLinkConfig1.m
@@ -40,8 +40,8 @@
 @implementation SoapActionsWANEthernetLinkConfig1
 
 
--(int)GetEthernetLinkStatusWithOutNewEthernetLinkStatus:(NSMutableString*)newethernetlinkstatus{
-    int ret = 0;
+-(NSInteger)GetEthernetLinkStatusWithOutNewEthernetLinkStatus:(NSMutableString*)newethernetlinkstatus{
+    NSInteger ret = 0;
 
     NSDictionary *parameters = nil;
     NSDictionary *output = nil;
