+++ src/upnp/UPnPDB.m
@@ -92,8 +92,8 @@
 }
 
 
--(int)addObserver:(UPnPDBObserver*)obs{
-    int ret = 0;
+-(NSUInteger)addObserver:(UPnPDBObserver*)obs{
+    NSUInteger ret = 0;
     [self lock];
     [mObservers addObject:obs];
     ret = [mObservers count];
@@ -101,8 +101,8 @@
     return ret;
 }
 
--(int)removeObserver:(UPnPDBObserver*)obs{
-    int ret = 0;
+-(NSUInteger)removeObserver:(UPnPDBObserver*)obs{
+    NSUInteger ret = 0;
     if ([mMutex tryLock]) {
         [mObservers removeObject:obs];
         ret = [mObservers count];
